feat: register scene load handler component
