+++ Assets/Scripts/ProjectLifetimeScope.cs
@@ -33,6 +33,7 @@ public class ProjectLifetimeScope : LifetimeScope
         builder.RegisterComponentInHierarchy<SelectionManager>();
         builder.RegisterComponentInHierarchy<PlayerManager>();
         builder.RegisterComponentInHierarchy<NetworkObjectInjector>();
+        builder.RegisterComponentInHierarchy<SceneLoadHandler>();
 
         builder.Register<UnitRegistry>(Lifetime.Singleton).As<IUnitRegistry>();
         builder.Register<PlayerCursorRegistry>(Lifetime.Singleton).As<IPlayerCursorRegistry>();
